+++ sdk/event/attempt/index.js
@@ -1,0 +1,5 @@
+const attempt = require('./attempt.js');
+
+exports.get = attempt.get;
+exports.query = attempt.query;
+exports.Attempt = attempt.Attempt;